pyproject.toml
[project]
name = "genai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.4",
    "langchain>=1.2.9",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
]
